added validation for bean

+++ smoothie_backend/src/main/resources/db/changelog/001_create_nutrition.sql
@@ -1,6 +1,6 @@
 CREATE TABLE nutrition (
    id serial PRIMARY KEY,
-   energy SMALLINT NOT NULL,
+   energy NUMERIC(5,2) NOT NULL,
    protein NUMERIC(5,2) NOT NULL,
    fat NUMERIC(5,2) NOT NULL,
    carbohydrate NUMERIC(5,2) NOT NULL
